switch from sigcontext to mcontext_t [#506]

+++ src/include/instremu.h
@@ -6,10 +6,10 @@ int instr_emu(sigcontext_t *scp, int pmode, int cnt);
 int decode_modify_segreg_insn(sigcontext_t *scp, int pmode,
     unsigned int *new_val);
 unsigned char instr_binary_byte(unsigned char op, unsigned char op1,
-                                   unsigned char op2, unsigned long *eflags);
+                                   unsigned char op2, unsigned *eflags);
 unsigned instr_binary_word(unsigned op, unsigned op1,
-                                   unsigned op2, unsigned long *eflags);
+                                   unsigned op2, unsigned *eflags);
 unsigned instr_binary_dword(unsigned op, unsigned op1,
-                                   unsigned op2, unsigned long *eflags);
+                                   unsigned op2, unsigned *eflags);
 
 #endif
